Update write_to_database methods
The string representations of level of theory objects used in the
database might differ from those used in AE and BAC classes. However,
write_to_database has to edit the database file as a text file instead
of treating it as a Python module, so make sure to match the correct
string representation of the level of theory in the database.