Call update_connectivity_values explicitly after desorbing an adsorbate.
When trying to estimate adsorbed CH3X, it creates a gas-phase CH3 and looks for the thermo.
It was not matching CH3 in a thermo library. Because the CH3 made from the CH3X had incorrect
connectivity values - i.e. the C atom had a connectivity1 attribute of 4, not 3.
The dummy_molecule.update() does not actually update the connectivity values.

In another commit, removing the bond and vertices from the graph will invalidate
and reset the connectivity values, which will in turn trigger a re-evaluation of them
during the atom sorting in update(), but it's safest to do it explicitly here.
Also, won't be redundant, because update() won't recalculate them.